Consolidate unix tools
