--- unix_integration/src/opt/cache_clear.rs
@@ -1,9 +1,0 @@
-#[derive(Debug, Parser)]
-struct CacheClearOpt {
-    #[clap(short, long)]
-    debug: bool,
-    #[clap(short, long, action = clap::ArgAction::SetTrue)]
-    version: bool,
-    #[clap(long)]
-    really: bool,
-}
