--- unix_integration/src/opt/cache_invalidate.rs
@@ -1,7 +1,0 @@
-#[derive(Debug, Parser)]
-struct CacheInvalidateOpt {
-    #[clap(short, long)]
-    debug: bool,
-    #[clap(short, long, action = clap::ArgAction::SetTrue)]
-    version: bool,
-}
