--- unix_integration/src/opt/unixd_status.rs
@@ -1,7 +1,0 @@
-#[derive(Debug, Parser)]
-struct UnixdStatusOpt {
-    #[clap(short, long)]
-    debug: bool,
-    #[clap(short, long, action = clap::ArgAction::SetTrue)]
-    version: bool,
-}
